requirements.txt
streamlit>=1.28.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.14.0
python-dotenv>=1.0.0
joblib>=1.3.0
folium>=0.15.0
streamlit-folium>=0.18.0
requests>=2.31.0
